requirements-tts.txt
espnet
parallel-wavegan
scipy
torchaudio
